minor refactor CosmJSOfflineSigner
